multimind-studio/app/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
pydantic-settings
python-dotenv
httpx
multimind-sdk
chromadb
pypdf
